+++ src/Graphics/GL/Skybox.cpp
@@ -234,7 +234,7 @@ GL::Skybox::Skybox()
 
 	m_vbo = std::make_unique<GL::Buffer>(ARRAY_BUFFER);
 	m_vbo->Bind();
-	m_vbo->Update(m_vertices.size() * sizeof(glm::vec3), &m_vertices[0]);
+	m_vbo->Update((unsigned int)m_vertices.size() * sizeof(glm::vec3), &m_vertices[0]);
 	
 	glGenVertexArrays(1, &m_vao);
 	glBindVertexArray(m_vao);
@@ -244,21 +244,21 @@ GL::Skybox::Skybox()
 
 	m_uvbo = std::make_unique<GL::Buffer>(ARRAY_BUFFER);
 	m_uvbo->Bind();
-	m_uvbo->Update(m_uvs.size() * sizeof(glm::vec2), &m_uvs[0]);
+	m_uvbo->Update((unsigned int)m_uvs.size() * sizeof(glm::vec2), &m_uvs[0]);
 
 	glEnableVertexAttribArray(1);
 	glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, 0, (void*)0);
 
 	m_nbo = std::make_unique<GL::Buffer>(ARRAY_BUFFER);
 	m_nbo->Bind();
-	m_nbo->Update(m_normals.size() * sizeof(glm::vec3), &m_normals[0]);
+	m_nbo->Update((unsigned int)m_normals.size() * sizeof(glm::vec3), &m_normals[0]);
 
 	glEnableVertexAttribArray(2);
 	glVertexAttribPointer(2, 3, GL_FLOAT, GL_FALSE, 0, (void*)0);
 
 	m_fbo = std::make_unique<GL::Buffer>(ELEMENT_ARRAY_BUFFER);
 	m_fbo->Bind();
-	m_fbo->Update(m_faces.size() * sizeof(std::uint32_t), &m_faces[0]);	
+	m_fbo->Update((unsigned int)m_faces.size() * sizeof(std::uint32_t), &m_faces[0]);
 }
 
 GL::Skybox::~Skybox()
